Allow toggling icomplete-vertical-mode in an active minibuffer
Before, if you toggled icomplete-vertical-mode from an active
minibuffer completion session, the necessary configuration was only
partially applied because part of it is added to the minibuffer setup
hook which wouldn't run until the next time the minibuffer was
entered. With this change the case of an active minibuffer is handled
specially and the hook is run immediately. This also requires undoing
the work of the setup hook if turning off the mode from an active
minibuffer (and this change also does that).