Fix compatibility with fido-mode
When using fido-mode, a built-in layer on top of icomplete-mode, the
latter is already enabled. Unfortunately, explicitly enabling
icomplete-mode (like done here) disables fido-mode, and so breaks the
user's setup. We instead need to enable icomplete-mode if and only if it
is not enabled. To be honest, this behavior of Emacs seems problematic,
and I'll report it upstream, but this fixes it here.